SCC: splits elements definition classes into different files
